stubs' children count serializtion modified
git-svn-id: http://svn.jetbrains.org/idea/Trunk/scala@21798 59f64563-22e8-0310-8d2c-98a15e1539f2

+++ scala_core/src/org/jetbrains/plugins/scala/lang/psi/stubs/elements/ScVariableElementType.scala
@@ -23,13 +23,13 @@ extends ScStubElementType[ScVariableStub, ScVariable](debugName) {
   def serialize(stub: ScVariableStub, dataStream: StubOutputStream): Unit = {
     dataStream.writeBoolean(stub.isDeclaration)
     val names = stub.getNames
-    dataStream.writeByte(names.length)
+    dataStream.writeInt(names.length)
     for (name <- names) dataStream.writeName(name)
   }
 
   def deserializeImpl(dataStream: StubInputStream, parentStub: Any): ScVariableStub = {
     val isDecl = dataStream.readBoolean
-    val namesLength = dataStream.readByte
+    val namesLength = dataStream.readInt
     val names = new Array[String](namesLength)
     for (i <- 0 to (namesLength - 1)) names(i) = StringRef.toString(dataStream.readName)
     val parent = parentStub.asInstanceOf[StubElement[PsiElement]]
